requirements.txt
Django>=3.0.3
djangorestframework>=3.11.0
django-crispy-forms>=1.8.1
Pillow>=7.0.0
django-bootstrap-pagination>=1.7.1
# sendgrid-django
# django-cookie-consent
python-dotenv

# Google Drive
# django-googledrive-storage

# AWS
# boto3>=1.11.7
# django_storages>=1.8

# Heroku
# dj-database-url>=0.5.0
# gunicorn>=19.9.0
# psycopg2>=2.7.2
whitenoise
